+++ src/main/java/edu/ie3/datamodel/io/factory/input/NodeAssetInputEntityData.java
@@ -8,14 +8,14 @@ package edu.ie3.datamodel.io.factory.input;
 import edu.ie3.datamodel.models.UniqueEntity;
 import edu.ie3.datamodel.models.input.NodeInput;
 import edu.ie3.datamodel.models.input.OperatorInput;
+import edu.ie3.datamodel.models.input.UniqueInputEntity;
 import java.util.Map;
 import java.util.Objects;
 
 /**
- * Data used by all factories used to create instances of {@link
- * edu.ie3.datamodel.models.input.InputEntity}s holding one {@link NodeInput} entity, thus needing
- * additional information about the {@link edu.ie3.datamodel.models.input.NodeInput}, which cannot
- * be provided through the attribute map.
+ * Data used by all factories used to create instances of {@link UniqueInputEntity}s holding one
+ * {@link NodeInput} entity, thus needing additional information about the {@link NodeInput}, which
+ * cannot be provided through the attribute map.
  */
 public class NodeAssetInputEntityData extends AssetInputEntityData {
   private final NodeInput node;