+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/IdCoordinateFactory.java
@@ -7,17 +7,16 @@ package edu.ie3.datamodel.io.factory.timeseries;
 
 import edu.ie3.datamodel.io.factory.Factory;
 import edu.ie3.datamodel.io.factory.SimpleFactoryData;
-import org.apache.commons.lang3.tuple.Pair;
-import org.locationtech.jts.geom.Point;
+import edu.ie3.datamodel.models.input.IdCoordinateInput;
 
 /**
  * Abstract class definition for a factory, that is able to build single mapping entries from
  * coordinate identifier to actual coordinate
  */
 public abstract class IdCoordinateFactory
-    extends Factory<Pair, SimpleFactoryData, Pair<Integer, Point>> {
+    extends Factory<IdCoordinateInput, SimpleFactoryData, IdCoordinateInput> {
   protected IdCoordinateFactory() {
-    super(Pair.class);
+    super(IdCoordinateInput.class);
   }
 
   /** @return the field id for the coordinate id */