+++ src/test/groovy/edu/ie3/datamodel/io/processor/input/InputEntityProcessorTest.groovy
@@ -521,7 +521,8 @@ class InputEntityProcessorTest extends Specification {
       "eStorage"   : "100.0",
       "eCons"      : "23.0",
       "sRated"     : "22.0",
-      "cosPhiRated": "0.9"
+      "cosPhiRated": "0.9",
+      "sRatedDC"     : "20.0",
     ]
 
     when: