+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvDataSourceTest.groovy
@@ -5,24 +5,14 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
-import edu.ie3.datamodel.exceptions.SourceException
 import edu.ie3.datamodel.io.naming.FileNamingStrategy
-import edu.ie3.datamodel.models.UniqueEntity
-import edu.ie3.datamodel.models.input.NodeInput
 import edu.ie3.datamodel.models.input.system.LoadInput
-import edu.ie3.test.common.SystemParticipantTestData as sptd
 import spock.lang.Shared
 import spock.lang.Specification
 
 import java.nio.file.Path
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.atomic.LongAdder
-import java.util.function.Function
-import java.util.stream.Collectors
 
 class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
-  @Shared
-  Function<Map<String, String>, String> uuidExtractor = { fieldToValues -> fieldToValues.get("uuid") }
 
   // Using a groovy bug to gain access to private methods in superclass:
   // by default, we cannot access private methods with parameters from abstract parent classes, introducing a
@@ -37,11 +27,6 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
     Map<String, String> buildFieldsToAttributes(
         final String csvRow, final String[] headline) {
       return super.buildFieldsToAttributes(csvRow, headline)
-    }
-
-    def <T extends UniqueEntity> Set<Map<String, String>> distinctRowsWithLog(
-        Class<T> entityClass, Collection<Map<String, String>> allRows) {
-      return super.distinctRowsWithLog(allRows, uuidExtractor, entityClass.simpleName, "UUID")
     }
 
     String[] parseCsvRow(
@@ -100,24 +85,7 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
     def path = Path.of("this/path/does-not-exist")
 
     expect:
-    dummyCsvSource.getSourceFields(() -> dummyCsvSource.connector.initReader(path)).isEmpty()
-  }
-
-  def "A CsvDataSource should throw an exception when retrieving column names for a non-readable CSV file"() {
-    given:
-    DummyCsvSource source = new DummyCsvSource(csvSep, participantsFolderPath, fileNamingStrategy)
-    def readerSupplier = () -> {
-      def reader = source.connector.initReader(LoadInput)
-      reader.close() // We simulate the file being unreadable by just closing the reader before passing it over
-      return reader
-    }
-
-    when:
-    source.getSourceFields(readerSupplier)
-
-    then:
-    def exc = thrown(SourceException)
-    exc.message.startsWith("Error while trying to read source")
+    dummyCsvSource.getSourceFields(path).isEmpty()
   }
 
   def "A CsvDataSource should build a valid fields to attributes map with valid data as expected"() {
@@ -311,106 +279,5 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8;25.0;100.0;0.95;98.0;test_bmTypeInput;50.0;25.0" || "wrong separator"
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput"           || "too less columns"
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput,,,,"       || "too much columns"
-  }
-
-  def "A CsvDataSource should collect be able to collect empty optionals when asked to do so"() {
-
-    given:
-    ConcurrentHashMap<Class<? extends UniqueEntity>, LongAdder> emptyCollector = new ConcurrentHashMap<>()
-    def nodeInputOptionals = [
-      Optional.of(sptd.hpInput.node),
-      Optional.empty(),
-      Optional.of(sptd.chpInput.node)
-    ]
-
-    when:
-    def resultingList = nodeInputOptionals.stream().filter(dummyCsvSource.isPresentCollectIfNot(NodeInput, emptyCollector)).collect(Collectors.toList())
-
-    then:
-    emptyCollector.size() == 1
-    emptyCollector.get(NodeInput).toInteger() == 1
-
-    resultingList.size() == 2
-    resultingList.get(0) == Optional.of(sptd.hpInput.node)
-    resultingList.get(1) == Optional.of(sptd.chpInput.node)
-  }
-
-  def "A CsvDataSource should return a given collection of csv row mappings as distinct rows collection correctly"() {
-
-    given:
-    def nodeInputRow = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z[UTC]",
-      "operates_from" : "2020-03-24T15:11:31Z[UTC]",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-
-    when:
-    def allRows = [nodeInputRow]* noOfEntities
-    def distinctRows = dummyCsvSource.distinctRowsWithLog(allRows, uuidExtractor, NodeInput.simpleName, "UUID")
-
-    then:
-    distinctRows.size() == distinctSize
-    distinctRows[0] == firstElement
-
-    where:
-    noOfEntities || distinctSize || firstElement
-    0            || 0            || null
-    10           || 1            || ["uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z[UTC]",
-      "operates_from" : "2020-03-24T15:11:31Z[UTC]",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"]
-  }
-
-  def "A CsvDataSource should return an empty set of csv row mappings if the provided collection of mappings contains duplicated UUIDs with different data"() {
-
-    given:
-    def nodeInputRow1 = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z[UTC]",
-      "operates_from" : "2020-03-24T15:11:31Z[UTC]",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-    def nodeInputRow2 = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_b",
-      "operates_until": "2020-03-25T15:11:31Z[UTC]",
-      "operates_from" : "2020-03-24T15:11:31Z[UTC]",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-
-    when:
-    def allRows = [nodeInputRow1, nodeInputRow2]* 10
-    def distinctRows = dummyCsvSource.distinctRowsWithLog(allRows, uuidExtractor, NodeInput.simpleName, "UUID")
-
-    then:
-    distinctRows.size() == 0
   }
 }