+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvTypeSourceTest.groovy
@@ -200,6 +200,7 @@ class CsvTypeSourceTest extends Specification implements CsvTestDataMeta {
       eCons == sptd.evTypeInput.eCons
       sRated == sptd.evTypeInput.sRated
       cosPhiRated == sptd.evTypeInput.cosPhiRated
+      sRatedDC == sptd.evTypeInput.sRatedDC
     }
   }
 }